+++ src/main/scala/com/codahale/jerkson/ser/EitherSerializer.scala
@@ -8,13 +8,13 @@ import org.codehaus.jackson.map.{SerializerProvider, JsonSerializer}
  * @author coda
  */
 class EitherSerializer extends JsonSerializer[Either[_, _]] {
-  def serialize(value: Either[_, _], json: JsonGenerator, provider: SerializerProvider) = {
+  def serialize(value: Either[_, _], json: JsonGenerator, provider: SerializerProvider) {
     val obj: Object = value match {
       case Left(o) => o.asInstanceOf[Object]
       case Right(o) => o.asInstanceOf[Object]
     }
 
-    val serializer = provider.findValueSerializer(obj.getClass)
+    val serializer = provider.findValueSerializer(obj.getClass, null)
     serializer.serialize(obj, json, provider)
   }
 }